seahuntr is actually 'Black Touch II' a Korean game that uses Shanghai II as a base.
