+++ src/mame/mame.lst
@@ -36466,9 +36466,6 @@ sdk86                           //
 seabattl                        // (c) 1980
 seabattla                       // (c) 1980
 
-@source:seahunter.cpp
-seahuntr
-
 @source:sealy.cpp
 crzyddz                         // (c) 2004? Sealy
 
@@ -37600,6 +37597,7 @@ shangha3up                      // (c) 1993 Sunsoft (US prototype from 'PCB001')
 kothello                        // (c) 1990 Success.
 shangha2                        // (c) 1989 Sunsoft (Sun Electronics)
 shangha2a                       // (c) 1989 Sunsoft (Sun Electronics)
+blktch2                        // bootleg
 shanghai                        // (c) 1988 Sunsoft (Sun Electronics)
 shanghaij                       // (c) 1988 Sunsoft (Sun Electronics)
